Further structure and precommit
At the moment the code structure is in place to extract data from XML files, but it doesn't yet return a valid
dataverifyr::ruleset() that can be written to YAML and used subsequently.

Some unittests are in place for individual functions but I'm very inexperienced at writing tests in R so these feel very
crude at the moment.

I've used `devtools`, `usethis::use_readme_rmd()` and `codemetar::write_codemeta()` to setup the package structure as
well as `precommit` to setup pre-commit hooks to ensure that good coding practices (as defined in [R
Packages](https://r-pkgs.org/)) are followed.