Adding R Package structure and precommit
At the moment the code structure is in place to extract data from XML files, but it doesn't yet return a valid
dataverifyr::ruleset() that can be written to YAML and used subsequently.

I've used `devtools`, `usethis::use_readme_rmd()` and `codemetar::write_codemeta()` to setup the package structure as
well as `precommit` to setup pre-commit hooks.